lotlist from int to long

+++ src/main/java/com/sericulture/marketandauction/model/api/marketauction/FarmerPaymentInfoRequestByLotList.java
@@ -11,5 +11,5 @@ import java.util.List;
 @EqualsAndHashCode
 public class FarmerPaymentInfoRequestByLotList extends FarmerPaymentInfoRequest{
 
-    private List<Integer> allottedLotList;
+    private List<Long> allottedLotList;
 }
